+++ PyPDF2_Fields/__init__.py
@@ -2,7 +2,7 @@ from .field_types import\
 	PdfFieldType,\
 	get_field_type
 
-from .pypdf2_fields import\
+from .misc import\
 	make_writer_from_reader,\
 	pair_fields_name_and_val,\
 	set_need_appearances,\
